--- lib/provider/ui/windows/solitaire/atom.ts
@@ -1,155 +1,0 @@
-import type { Card, Place } from "./type";
-
-import { atom } from "jotai";
-
-import { atomWithWriteOnly } from "~/lib/atom-with-write-only";
-
-export const foundationAtom = atom<Record<number, Card[]>>({
-  0: [],
-  1: [],
-  2: [],
-  3: [],
-});
-export const wasteAtom = atom<Card[]>([]);
-export const tableauAtom = atom<Record<number, Card[]>>({});
-
-interface MoveCardParams {
-  card: Card;
-  from: {
-    place: Place;
-    column: number;
-  };
-  to: {
-    place: Place;
-    column: number;
-  };
-}
-
-export const moveCardAtom = atomWithWriteOnly(
-  (get, set, { card, from, to }: MoveCardParams) => {
-    switch (from.place) {
-      case "tableau": {
-        const tableau = { ...get(tableauAtom) };
-
-        if (to.place === "tableau") {
-          const targetColumn = tableau[to.column];
-
-          if (!canMoveToTableau(card, targetColumn)) return;
-
-          const sourceColumn = tableau[from.column];
-          const cardIndex = sourceColumn.findIndex((c) => c.id === card.id);
-          const cardsToMove = sourceColumn.splice(cardIndex);
-
-          for (const c of cardsToMove) {
-            c.place = to.place;
-            c.column = to.column;
-          }
-
-          const lastCard = sourceColumn.at(-1);
-          if (lastCard) lastCard.facingUp = true;
-
-          tableau[to.column].push(...cardsToMove);
-          set(tableauAtom, tableau);
-        } else if (to.place === "foundation") {
-          const foundation = { ...get(foundationAtom) };
-          const targetColumn = foundation[to.column];
-
-          if (!canMoveToFoundation(card, targetColumn)) return;
-
-          const sourceColumn = tableau[from.column];
-          const cardIndex = sourceColumn.findIndex((c) => c.id === card.id);
-          const cardsToMove = sourceColumn.splice(cardIndex);
-
-          for (const c of cardsToMove) {
-            c.place = to.place;
-            c.column = to.column;
-          }
-
-          const lastCard = sourceColumn.at(-1);
-          if (lastCard) lastCard.facingUp = true;
-
-          foundation[to.column].push(card);
-          set(foundationAtom, foundation);
-          set(tableauAtom, tableau);
-        }
-
-        break;
-      }
-      case "waste": {
-        const waste = [...get(wasteAtom)];
-        const cardIndex = waste.findIndex((c) => c.id === card.id);
-        const [movedCard] = waste.splice(cardIndex, 1);
-
-        if (to.place === "tableau") {
-          const tableau = { ...get(tableauAtom) };
-          const targetColumn = tableau[to.column];
-
-          if (!canMoveToTableau(movedCard, targetColumn)) return;
-
-          movedCard.place = to.place;
-          movedCard.column = to.column;
-          set(wasteAtom, waste);
-
-          tableau[to.column].push(movedCard);
-          set(tableauAtom, tableau);
-        } else if (to.place === "foundation") {
-          const foundation = { ...get(foundationAtom) };
-          const targetColumn = foundation[to.column];
-
-          if (!canMoveToFoundation(movedCard, targetColumn)) return;
-
-          movedCard.place = to.place;
-          movedCard.column = to.column;
-          set(wasteAtom, waste);
-
-          foundation[to.column].push(movedCard);
-          set(foundationAtom, foundation);
-        }
-
-        break;
-      }
-      case "foundation": {
-        if (to.place === "tableau") {
-          const foundation = { ...get(foundationAtom) };
-          const [movedCard] = foundation[from.column].splice(-1, 1);
-          const tableau = { ...get(tableauAtom) };
-          const targetColumn = tableau[to.column];
-
-          if (!canMoveToTableau(movedCard, targetColumn)) return;
-
-          movedCard.place = to.place;
-          movedCard.column = to.column;
-
-          set(foundationAtom, foundation);
-
-          tableau[to.column].push(movedCard);
-          set(tableauAtom, tableau);
-        }
-
-        break;
-      }
-    }
-  },
-);
-
-function canMoveToFoundation(sourceCard: Card, targetColumn: Card[]): boolean {
-  const lastCard = targetColumn.at(-1);
-
-  if (!lastCard) return sourceCard.number === 1;
-  console.log(sourceCard, lastCard);
-  return (
-    lastCard.type === sourceCard.type &&
-    lastCard.number === sourceCard.number - 1
-  );
-}
-
-function canMoveToTableau(sourceCard: Card, targetColumn: Card[]): boolean {
-  const lastCard = targetColumn.at(-1);
-
-  if (!lastCard) return sourceCard.number === 13;
-
-  return (
-    lastCard.black !== sourceCard.black &&
-    lastCard.number === sourceCard.number + 1
-  );
-}
